% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRpiechart.R
\name{LRpiechart}
\alias{LRpiechart}
\title{LRpiechart}
\usage{
LRpiechart(dat, colnum = "celltype1", colors = col_vector, main = NULL)
}
\arguments{
\item{dat}{A data frame containing the data to be plotted.}

\item{colnum}{A string specifying the column name to be used for the pie chart.}

\item{colors}{A vector of colors to be used for the pie chart segments.}

\item{main}{A string specifying the title of the pie chart.}
}
\value{
A ggplot object representing the pie chart.
}
\description{
Plot a pie chart for a given column in the data frame.
}
