% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_clustering.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA Results}
\usage{
plot_pca(
  pca_coords,
  clusters,
  pc_x = 1,
  pc_y = 2,
  color_by = NULL,
  colors = NULL,
  point_size = 3,
  main = NULL
)
}
\arguments{
\item{pca_coords}{Matrix of PCA coordinates (samples x PCs)}

\item{clusters}{Vector of cluster assignments}

\item{pc_x}{Principal component for x-axis (default: 1)}

\item{pc_y}{Principal component for y-axis (default: 2)}

\item{color_by}{Vector to color points by. If NULL, uses clusters (default: NULL)}

\item{colors}{Color palette. If NULL, uses default colors (default: NULL)}

\item{point_size}{Size of points (default: 3)}

\item{main}{Title for the plot (default: NULL)}
}
\value{
A ggplot object
}
\description{
Creates scatter plots of PCA results colored by clusters or other metadata.
}
\examples{
\dontrun{
# Perform clustering
results <- dimensionality_clustering(lr_matrix)

# Plot PCA colored by clusters
plot_pca(results$pca_coords, results$clusters, pc_x = 1, pc_y = 2)
}

}
